Refactor: 클래스가 이동됨에 따라 import 수정
Related to: #47
