+++ src/test/java/pulleydoreurae/careerquestbackend/certification/service/ReviewServiceTest.java
@@ -21,8 +21,8 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
-import pulleydoreurae.careerquestbackend.certification.domain.dto.ReviewRequest;
-import pulleydoreurae.careerquestbackend.certification.domain.dto.ReviewResponse;
+import pulleydoreurae.careerquestbackend.certification.domain.dto.request.ReviewRequest;
+import pulleydoreurae.careerquestbackend.certification.domain.dto.response.ReviewResponse;
 import pulleydoreurae.careerquestbackend.certification.domain.entity.Review;
 import pulleydoreurae.careerquestbackend.certification.domain.entity.ReviewViewCheck;
 import pulleydoreurae.careerquestbackend.certification.repository.ReviewLikeRepository;
